Ajout des widget et correction des bugs terminé
