+++ app.json
@@ -18,8 +18,8 @@
         },
         "android": {
             "adaptiveIcon": {
-                "foregroundImage": "./assets/adaptive-icon.png",
-                "backgroundColor": "#ffffff"
+                "foregroundImage": "./assets/icon.png",
+                "backgroundColor": "#efecca"
             },
             "package": "com.gaia.natureskitchen"
         },
