+++ src/domain/interfaces/props/widgets/small-multiple-intakes-props.ts
@@ -1,4 +1,3 @@
-import { NutrientData } from '~/domain/interfaces/props/nutrient-data';
 import { NutrientsEnum } from '~/domain/interfaces/enum/nutrients-enum';
 
 export type SmallMultipleIntakesProps = { nutrients: NutrientsEnum[] };
