+++ src/infrastructure/ui/pages/home-page/home-page.tsx
@@ -6,17 +6,9 @@ import HomePageBasket from '~/infrastructure/ui/pages/home-page/component/backgr
 import HomePageStyle from '~/infrastructure/ui/pages/home-page/home-page-style';
 import HomePageBlobsTop from '~/infrastructure/ui/pages/home-page/component/background/home-page-blobs-top';
 import useHomePageData from '~/infrastructure/ui/pages/home-page/hooks';
-import EcoScore from '~/infrastructure/ui/shared/component/widgets/eco-score/widget-ecoscore';
-import LargeIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/large/large-intakes';
-import { WidgetEnum } from '~/domain/interfaces/enum/widget-enum';
-import WidgetAnecdote from '~/infrastructure/ui/shared/component/widgets/anecdote/widget-anecdote';
-import SmallBasicIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/small-basic/small-basic-intakes';
-import SmallMultipleIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/small-multiple/small-multiple-intakes';
-import WidgetCalorie from '~/infrastructure/ui/shared/component/widgets/calorie/widget-calorie';
-import { NutrientsEnum } from '~/domain/interfaces/enum/nutrients-enum';
 
 const HomePage = () => {
-    const { username, chooseRightDynamicImage, widgetsParams } = useHomePageData();
+    const { username, chooseRightDynamicImage, widgetField } = useHomePageData();
 
     return (
         <View>
@@ -35,67 +27,7 @@ const HomePage = () => {
 
                 <HomePageBasket asset={chooseRightDynamicImage()} />
 
-                <View style={HomePageStyle.widgetContainer}>
-                    <View style={HomePageStyle.widgetContainerFirstRow}>
-                        {widgetsParams.line1.map((widget, index) => {
-                            console.log('line1 : ', widget);
-                            switch (widget.type) {
-                                case WidgetEnum.Anecdote:
-                                    return <WidgetAnecdote key={index} />;
-                                case WidgetEnum.EcoScore:
-                                    return <EcoScore key={index} />;
-                                case WidgetEnum.Large:
-                                    return (
-                                        <LargeIntakes
-                                            key={index}
-                                            nutrients={[
-                                                NutrientsEnum.Protein,
-                                                NutrientsEnum.Lipid,
-                                                NutrientsEnum.Carbohydrate
-                                            ]}
-                                        />
-                                    );
-                                case WidgetEnum.SmallBasic:
-                                    return <SmallBasicIntakes key={index} nutrient={widget.nutrient} />;
-                                case WidgetEnum.SmallMultiple:
-                                    return <SmallMultipleIntakes key={index} nutrients={widget.nutrients} />;
-                                case WidgetEnum.Calorie:
-                                    return <WidgetCalorie key={index} />;
-                                default:
-                            }
-                        })}
-                    </View>
-
-                    <View style={HomePageStyle.widgetContainerTwoWidgetRow}>
-                        {widgetsParams.line2.map((widget, index) => {
-                            console.log('line2 : ', widget);
-                            switch (widget.type) {
-                                case WidgetEnum.Anecdote:
-                                    return <WidgetAnecdote key={index} />;
-                                case WidgetEnum.EcoScore:
-                                    return <EcoScore key={index} />;
-                                case WidgetEnum.Large:
-                                    return (
-                                        <LargeIntakes
-                                            key={index}
-                                            nutrients={[
-                                                NutrientsEnum.Protein,
-                                                NutrientsEnum.Lipid,
-                                                NutrientsEnum.Carbohydrate
-                                            ]}
-                                        />
-                                    );
-                                case WidgetEnum.SmallBasic:
-                                    return <SmallBasicIntakes key={index} nutrient={widget.nutrient} />;
-                                case WidgetEnum.SmallMultiple:
-                                    return <SmallMultipleIntakes key={index} nutrients={widget.nutrients} />;
-                                case WidgetEnum.Calorie:
-                                    return <WidgetCalorie key={index} />;
-                                default:
-                            }
-                        })}
-                    </View>
-                </View>
+                <View style={HomePageStyle.widgetContainer}>{widgetField}</View>
             </View>
         </View>
     );
