+++ src/infrastructure/ui/pages/widget-page/widget-page.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import { ScrollView, Text, TouchableOpacity, View } from 'react-native';
 import { observer } from 'mobx-react';
 import WidgetPageStyle from '~/infrastructure/ui/pages/widget-page/widget-page-style';
@@ -6,193 +6,35 @@ import LoginPageTreeClassicLogo from '~/infrastructure/ui/pages/settings-page/co
 import GenericBackArrowIcon from '~/infrastructure/ui/shared/component/generic-back-arrow-icon/generic-back-arrow-icon';
 import WidgetPageBlobsTop from '~/infrastructure/ui/pages/widget-page/component/background/widget-page-blobs-top';
 import WidgetPageBlobsBottom from '~/infrastructure/ui/pages/widget-page/component/background/widget-page-blobs-bottom';
-import { useStore } from '~/infrastructure/controllers/store';
 import SmallBlankWidget from '~/infrastructure/ui/pages/widget-page/component/small-widget/small-blank-widget';
 import GenericHeaderText from '~/infrastructure/ui/shared/component/texts/generic-header-text/generic-header-text';
 import LargeBlankWidget from '~/infrastructure/ui/pages/widget-page/component/large-widget/large-blank-widget';
-import WidgetSlot from '~/infrastructure/ui/pages/widget-page/component/widget-slot/widget-slot';
 import GenericButton from '~/infrastructure/ui/shared/component/generic-button/generic-button';
-import { ColorEnum } from '~/domain/interfaces/enum/color-enum';
-import { WidgetsParams } from '~/domain/interfaces/props/widgets/widgets-params';
 import { WidgetEnum } from '~/domain/interfaces/enum/widget-enum';
-import WidgetAnecdote from '~/infrastructure/ui/shared/component/widgets/anecdote/widget-anecdote';
-import EcoScore from '~/infrastructure/ui/shared/component/widgets/eco-score/widget-ecoscore';
-import LargeIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/large/large-intakes';
-import SmallBasicIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/small-basic/small-basic-intakes';
-import SmallMultipleIntakes from '~/infrastructure/ui/shared/component/widgets/my-intakes/small-multiple/small-multiple-intakes';
-import { NutrientsEnum } from '~/domain/interfaces/enum/nutrients-enum';
 import CustomModal from '~/infrastructure/ui/shared/component/modal/custom-modal';
 import CustomSvg from '~/infrastructure/ui/shared/custom-svg';
-import WidgetCalorie from '~/infrastructure/ui/shared/component/widgets/calorie/widget-calorie';
 import GenericDropDown from '~/infrastructure/ui/shared/component/inputs/generic-dropdown/generic-dropdown';
-import { WidgetItem } from '~/domain/interfaces/props/widgets/widget-item';
+import useWidgetPageData from '~/infrastructure/ui/pages/widget-page/hooks';
 
-// eslint-disable-next-line max-lines-per-function
 const WidgetPage = () => {
     const {
-        NavigationStore: { goBack },
-        DataStore: { setWidgetParams }
-    } = useStore();
-
-    const confirmButtonStyle = {
-        container: {
-            backgroundColor: ColorEnum.ClassicGreen,
-            borderRadius: 20,
-            width: 105,
-            height: 45
-        },
-        text: {
-            color: ColorEnum.ClassicGrey,
-            fontSize: 18,
-            fontWeight: 700
-        }
-    };
-
-    const [handleDrop, setHandleDrop] = useState<{ isLine: boolean; id: number }>();
-
-    const [widgetDropped, setWidgetDropped] = useState<{ type: 'small' | 'large'; x: number; y: number } | undefined>(
-        undefined
-    );
-
-    const [isAddSmallWidgetModalOpen, setIsAddSmallWidgetModalOpen] = useState(false);
-
-    const [newWidgetParams, setNewWidgetParams] = useState<WidgetsParams>({
-        line1: [{ type: WidgetEnum.Slot }, { type: WidgetEnum.Slot }],
-        line2: [{ type: WidgetEnum.Slot }, { type: WidgetEnum.Slot }]
-    });
-
-    useEffect(() => {
-        if (handleDrop) {
-            if (handleDrop.isLine) {
-                setNewWidgetParams((prevState) => {
-                    return {
-                        line1: handleDrop.id === 1 ? [{ type: WidgetEnum.Large }] : [...prevState.line1],
-                        line2: handleDrop.id === 2 ? [{ type: WidgetEnum.Large }] : [...prevState.line2]
-                    };
-                });
-            } else {
-                console.log('open');
-                setIsAddSmallWidgetModalOpen(true);
-            }
-        }
-    }, [handleDrop]);
-
-    const [choosenWidget, setChoosenWidget] = useState<WidgetEnum>(WidgetEnum.Slot);
-
-    const toggleModal = useCallback((value: boolean) => {
-        setIsAddSmallWidgetModalOpen(value);
-        setChoosenWidget(WidgetEnum.Slot);
-    }, []);
-
-    const nutrientsOptions: { label: string; value: string }[] = [
-        { label: NutrientsEnum.Sugar, value: NutrientsEnum.Sugar },
-        { label: NutrientsEnum.Lipid, value: NutrientsEnum.Lipid },
-        { label: NutrientsEnum.Carbohydrate, value: NutrientsEnum.Carbohydrate },
-        { label: NutrientsEnum.FattyAcid, value: NutrientsEnum.FattyAcid },
-        { label: NutrientsEnum.Fiber, value: NutrientsEnum.Fiber },
-        { label: NutrientsEnum.Salt, value: NutrientsEnum.Salt },
-        { label: NutrientsEnum.Protein, value: NutrientsEnum.Protein }
-    ];
-
-    const [firstNutrient, setFirstNutrient] = useState<{ label: string; value: string }>({
-        label: NutrientsEnum.Sugar,
-        value: NutrientsEnum.Sugar
-    });
-
-    const [secondNutrient, setSecondNutrient] = useState<{ label: string; value: string }>({
-        label: NutrientsEnum.Sugar,
-        value: NutrientsEnum.Sugar
-    });
-
-    const [thirdNutrient, setThirdNutrient] = useState<{ label: string; value: string }>({
-        label: NutrientsEnum.Sugar,
-        value: NutrientsEnum.Sugar
-    });
-
-    const getNewWidgetParamsWithSmallModal = (prev: WidgetsParams, id: number, item: WidgetItem): WidgetsParams => {
-        switch (id) {
-            case 1:
-                return {
-                    line1: [{ ...item }, prev.line1[1]],
-                    line2: [...prev.line2]
-                };
-            case 2:
-                return {
-                    line1: [prev.line1[0], { ...item }],
-                    line2: [...prev.line2]
-                };
-            case 3:
-                return {
-                    line1: [...prev.line1],
-                    line2: [{ ...item }, prev.line2[1]]
-                };
-            case 4:
-                return {
-                    line1: [...prev.line1],
-                    line2: [prev.line2[0], { ...item }]
-                };
-        }
-    };
-
-    const handleModalConfirm = useCallback(() => {
-        switch (choosenWidget) {
-            case WidgetEnum.Anecdote:
-                setNewWidgetParams((prevState) => {
-                    return getNewWidgetParamsWithSmallModal(prevState, handleDrop.id, { type: WidgetEnum.Anecdote });
-                });
-                break;
-            case WidgetEnum.EcoScore:
-                setNewWidgetParams((prevState) => {
-                    return getNewWidgetParamsWithSmallModal(prevState, handleDrop.id, { type: WidgetEnum.EcoScore });
-                });
-                break;
-            case WidgetEnum.Calorie:
-                setNewWidgetParams((prevState) => {
-                    return getNewWidgetParamsWithSmallModal(prevState, handleDrop.id, { type: WidgetEnum.Calorie });
-                });
-                break;
-            case WidgetEnum.SmallBasic:
-                setNewWidgetParams((prevState) => {
-                    return getNewWidgetParamsWithSmallModal(prevState, handleDrop.id, {
-                        type: WidgetEnum.SmallBasic,
-                        nutrient: firstNutrient.value as NutrientsEnum
-                    });
-                });
-                break;
-            case WidgetEnum.SmallMultiple:
-                setNewWidgetParams((prevState) => {
-                    return getNewWidgetParamsWithSmallModal(prevState, handleDrop.id, {
-                        type: WidgetEnum.SmallMultiple,
-                        nutrients: [
-                            firstNutrient.value as NutrientsEnum,
-                            secondNutrient.value as NutrientsEnum,
-                            thirdNutrient.value as NutrientsEnum
-                        ]
-                    });
-                });
-        }
-        setIsAddSmallWidgetModalOpen(false);
-    }, [choosenWidget, firstNutrient.value, secondNutrient.value, thirdNutrient.value]);
-
-    const [isErrorModal, setIsErrorModal] = useState(false);
-
-    const handlePageConfig = useCallback(() => {
-        let isEntryValid = true;
-        newWidgetParams.line1.forEach((widget) => {
-            if (widget.type === WidgetEnum.Slot) isEntryValid = false;
-        });
-        newWidgetParams.line2.forEach((widget) => {
-            if (widget.type === WidgetEnum.Slot) isEntryValid = false;
-        });
-
-        if (isEntryValid) {
-            setWidgetParams(newWidgetParams);
-            goBack();
-        } else {
-            setIsErrorModal(true);
-        }
-    }, [goBack, newWidgetParams, setWidgetParams]);
-
+        setWidgetDropped,
+        handlePageConfig,
+        isErrorModal,
+        setIsErrorModal,
+        handleModalConfirm,
+        isAddSmallWidgetModalOpen,
+        toggleModal,
+        confirmButtonStyle,
+        setSecondNutrient,
+        setFirstNutrient,
+        nutrientsOptions,
+        setThirdNutrient,
+        goBack,
+        setChoosenWidget,
+        choosenWidget,
+        widgetField
+    } = useWidgetPageData();
     return (
         <View style={WidgetPageStyle.container}>
             <View style={WidgetPageStyle.background}>
@@ -212,80 +54,7 @@ const WidgetPage = () => {
                     <LargeBlankWidget setWidgetDropped={setWidgetDropped} />
                 </View>
                 <Text style={WidgetPageStyle.instructions}>Glissez vos widgets ci-dessous</Text>
-                <View style={WidgetPageStyle.widgetArrival}>
-                    <View style={WidgetPageStyle.widgetArrivalRow}>
-                        {newWidgetParams.line1.map((widget, index) => {
-                            switch (widget.type) {
-                                case WidgetEnum.Anecdote:
-                                    return <WidgetAnecdote key={index} />;
-                                case WidgetEnum.EcoScore:
-                                    return <EcoScore key={index} />;
-                                case WidgetEnum.Large:
-                                    return (
-                                        <LargeIntakes
-                                            key={index}
-                                            nutrients={[
-                                                NutrientsEnum.Protein,
-                                                NutrientsEnum.Lipid,
-                                                NutrientsEnum.Carbohydrate
-                                            ]}
-                                        />
-                                    );
-                                case WidgetEnum.SmallBasic:
-                                    return <SmallBasicIntakes key={index} nutrient={widget.nutrient} />;
-                                case WidgetEnum.SmallMultiple:
-                                    return <SmallMultipleIntakes key={index} nutrients={widget.nutrients} />;
-                                case WidgetEnum.Calorie:
-                                    return <WidgetCalorie key={index} />;
-                                default:
-                                    return (
-                                        <WidgetSlot
-                                            key={index}
-                                            id={index + 1}
-                                            setHandleDrop={setHandleDrop}
-                                            widgetDropped={widgetDropped}
-                                        />
-                                    );
-                            }
-                        })}
-                    </View>
-                    <View style={WidgetPageStyle.widgetArrivalRow}>
-                        {newWidgetParams.line2.map((widget, index) => {
-                            switch (widget.type) {
-                                case WidgetEnum.Anecdote:
-                                    return <WidgetAnecdote key={index} />;
-                                case WidgetEnum.EcoScore:
-                                    return <EcoScore key={index} />;
-                                case WidgetEnum.Large:
-                                    return (
-                                        <LargeIntakes
-                                            key={index + 2}
-                                            nutrients={[
-                                                NutrientsEnum.Protein,
-                                                NutrientsEnum.Lipid,
-                                                NutrientsEnum.Carbohydrate
-                                            ]}
-                                        />
-                                    );
-                                case WidgetEnum.SmallBasic:
-                                    return <SmallBasicIntakes key={index} nutrient={widget.nutrient} />;
-                                case WidgetEnum.SmallMultiple:
-                                    return <SmallMultipleIntakes key={index} nutrients={widget.nutrients} />;
-                                case WidgetEnum.Calorie:
-                                    return <WidgetCalorie key={index} />;
-                                default:
-                                    return (
-                                        <WidgetSlot
-                                            key={index}
-                                            id={index + 3}
-                                            setHandleDrop={setHandleDrop}
-                                            widgetDropped={widgetDropped}
-                                        />
-                                    );
-                            }
-                        })}
-                    </View>
-                </View>
+                <View style={WidgetPageStyle.widgetArrival}>{widgetField}</View>
                 <View style={WidgetPageStyle.footer}>
                     <GenericButton title="Confirmer" onPress={handlePageConfig} style={confirmButtonStyle} />
                 </View>
