+++ src/infrastructure/ui/shared/component/widgets/anecdote/hooks.ts
@@ -1,9 +1,10 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import useEffectOnce from '~/infrastructure/ui/shared/helper/use-effect-once';
 import { anecdotesObject } from '~/domain/entities/constants/anecdote-constants';
 import getRandomNumberInArrayLength from '~/infrastructure/ui/shared/helper/get-random-number-in-array-length';
 
 const useWidgetAnecdoteData = () => {
+    /* eslint-disable-next-line */
     const [annecdoteObject, setAnnecdoteObject] = useState<{ text: string; icon: any }>();
 
     useEffectOnce(() => {
