+++ src/infrastructure/ui/shared/component/widgets/my-intakes/small-basic/small-basic-intakes.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Text, View } from 'react-native';
 import { AnimatedCircularProgress } from 'react-native-circular-progress';
 import CustomFontInterBold from '~/application/utils/font/custom-font-inter-bold';
-import { NutrientsEnum } from '~/domain/interfaces/enum/nutrients-enum';
 import useSmallBasicIntakesData from '~/infrastructure/ui/shared/component/widgets/my-intakes/small-basic/hooks';
 import { ColorEnum } from '~/domain/interfaces/enum/color-enum';
 import { SmallBasicIntakesProps } from '~/domain/interfaces/props/widgets/small-basic-intakes-props';
